refactor name persister structure
